==> programs/15_table5_webstats.R <==
# Tabulate statistics and make graphs for the AEA data editor report
# Harry Son
# 12/11/2020

# Inputs
#   - file.path(jiraanon,"jira.anon.RDS") 
# Outputs


### Load libraries 
### Requirements: have library *here*
source(here::here("global-libraries.R"),echo=TRUE)
source(here::here("programs","config.R"),echo=TRUE)
library(stargazer)

# read back in

utilizationReport <- readRDS(file=file.path(icpsrbase,"anonUtilizationReport.Rds"))

icpsr.stats.mb <- readRDS(file=file.path(icpsrbase,"icpsr.stats.mb.Rds"))
icpsr.stats.gb <- readRDS(file=file.path(icpsrbase,"icpsr.stats.gb.Rds"))

webstats <- utilizationReport %>%
  summarize(totalDownloads = sum(Downloads),
            totalViews     = sum(Views),
            totalPublished = sum(is_published),
            maxDownloads   = max(Downloads),
            medianDownloads= median(Downloads))

update_latexnums("icpsrutilization",icpsr.utilization.date)
update_latexnums("icpsrtotalDownloads",webstats$totalDownloads)
update_latexnums("icpsrtotalDownloadsHT",round(webstats$totalDownloads/100000,0))
update_latexnums("icpsrmaxDownloads",webstats$maxDownloads)
update_latexnums("icpsrmedianDownloads",webstats$medianDownloads)
update_latexnums("icpsrtotalViews",webstats$totalViews)
update_latexnums("icpsrtotalViewsM",round(webstats$totalViews/1000000,1))
update_latexnums("icpsrtotalPublished",webstats$totalPublished)



# output table

webstats %>% 
  bind_cols(icpsr.stats.gb) %>%
  mutate(Repository = "ICPSR",
         Published   = prettyNum(totalPublished,big.mark = ","),
         `Downloads` = prettyNum(totalDownloads,big.mark = ","),
         `Views`     = prettyNum(totalViews,big.mark = ","),
         Uploads     = prettyNum(n,big.mark = ","),
         `Files`     = prettyNum(files,big.mark = ","),
         `Size (GB)` = prettyNum(sum,big.mark = ",")) %>%
  select(Repository,Published,Downloads,Views,Uploads,Files,`Size (GB)`)->
  icpsr.stats.table


# Do the same for Zenodo, though we only output latexnums
# The data were generated by hand (this year), running
# python3 zenodo_pull.py
# Requirements: requests
# dbl  (8): id, revision, downloads, views, unique_downloads, unique_views, 
#
# dbl (2): sumsize, filecount
# The number of uploads was hand-counted: 4


utilizationReport.z <- read_csv(file.path(zenodobase,paste0("zenodo_data_",report.year,".csv"))) %>%
  mutate(date_created = as.Date(substr(created, 1,10), "%Y-%m-%d")) 

sumstats.z          <- read_csv(file.path(zenodobase,paste0("zenodo_data_",report.year,"_summary.csv"))) %>%
  rename(files = filecount, sum=sumsize) %>%
  mutate(sum = sum/1024/1024/1024)

webstats <- utilizationReport.z %>%
  mutate(this.year = ( date_created >= firstday & date_created <= lastday)) %>%
  summarize(totalDownloads = sum(unique_downloads),
            totalViews     = sum(unique_views),
            totalPublished = n(),
            maxDownloads   = max(unique_downloads),
            medianDownloads= median(unique_downloads),
            uploads        = sum(this.year)) %>%
  bind_cols(sumstats.z)

update_latexnums("zenodototalDownloads",webstats$totalDownloads)
update_latexnums("zenodototalDownloadsHT",round(webstats$totalDownloads/100000,0))
update_latexnums("zenodomaxDownloads",webstats$maxDownloads)
update_latexnums("zenodomedianDownloads",webstats$medianDownloads)
update_latexnums("zenodototalViews",webstats$totalViews)
update_latexnums("zenodototalViewsM",round(webstats$totalViews/1000000,1))
update_latexnums("zenodototalPublished",webstats$totalPublished)
update_latexnums("zenodototalSizeGB",round(webstats$sum,0))
update_latexnums("zenodototalSizeTB",round(webstats$sum/1024,1))
update_latexnums("zenodototalFiles",webstats$files)

webstats %>% 
  mutate(Repository = "Zenodo",
          Published   = prettyNum(totalPublished,big.mark = ","),
         `Downloads` = prettyNum(totalDownloads,big.mark = ","),
         `Views`     = prettyNum(totalViews,big.mark = ","),
         Uploads     = prettyNum(uploads,big.mark = ","),
         `Files`     = prettyNum(files,big.mark = ","),
         `Size (GB)` = prettyNum(round(sum,2),big.mark = ",")) %>%
  select(Repository,Published,Downloads,Views,Uploads,Files,`Size (GB)`)->
  zenodo.stats.table

# combine

stats.table <- bind_rows(icpsr.stats.table, zenodo.stats.table)

stargazer(stats.table,style = "aer",
          summary = FALSE,
          out = file.path(tables,"n_webstats.tex"),
          out.header = FALSE,
          float = FALSE,
          rownames = FALSE,
          digit.separator = ",",
          digit.separate = 3
)
#
